# docs/swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for Bisky",
        "title": "Bisky API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/metadata/discover": {
            "get": {
                "description": "Get homepage-style discover feeds from provider (anilist recommended)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "metadata"
                ],
                "summary": "Metadata discover",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Provider type: anilist|anidb|tvdb (default anidb)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit per section",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/metadata.DiscoverResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httperr.APIErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httperr.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/metadata/episodes/{externalId}": {
            "get": {
                "description": "List episodes metadata by provider external id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "metadata"
                ],
                "summary": "Metadata list episodes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Provider external id",
                        "name": "externalId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Provider type: anidb|tvdb (default anidb)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/metadata.EpisodeResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httperr.APIErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httperr.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/metadata/search": {
            "get": {
                "description": "Search metadata by provider type (anidb default, tvdb optional)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "metadata"
                ],
                "summary": "Metadata search",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "query",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Provider type: anidb|tvdb (default anidb)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/metadata.SearchHitResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httperr.APIErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httperr.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/metadata/show/{externalId}": {
            "get": {
                "description": "Get show metadata by provider external id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "metadata"
                ],
                "summary": "Metadata get show",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Provider external id",
                        "name": "externalId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Provider type: anidb|tvdb (default anidb)",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/metadata.ShowResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httperr.APIErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httperr.APIErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Fetch provider show by external id and create a local show record",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "metadata"
                ],
                "summary": "Metadata add show",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Provider external id",
                        "name": "externalId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Provider type: anidb|anilist|tvdb (default anidb)",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/metadata.AddShowResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httperr.APIErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httperr.APIErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "httperr.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "httperr.APIErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/httperr.APIError"
                }
            }
        },
        "metadata.AddShowResponse": {
            "type": "object",
            "properties": {
                "altTitles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "bannerUrl": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "endDate": {
                    "type": "string"
                },
                "episodeCount": {
                    "type": "integer"
                },
                "externalId": {
                    "type": "string"
                },
                "internalShowId": {
                    "type": "string"
                },
                "posterUrl": {
                    "type": "string"
                },
                "seasonCount": {
                    "type": "integer"
                },
                "startDate": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "synopsis": {
                    "type": "string"
                },
                "titleOriginal": {
                    "type": "string"
                },
                "titlePreferred": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "metadata.DiscoverResponse": {
            "type": "object",
            "properties": {
                "currentlyAiring": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/metadata.ShowResponse"
                    }
                },
                "popular": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/metadata.ShowResponse"
                    }
                },
                "topRated": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/metadata.ShowResponse"
                    }
                },
                "trending": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/metadata.ShowResponse"
                    }
                },
                "upcoming": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/metadata.ShowResponse"
                    }
                }
            }
        },
        "metadata.EpisodeResponse": {
            "type": "object",
            "properties": {
                "airDate": {
                    "type": "string"
                },
                "episodeNumber": {
                    "type": "integer"
                },
                "externalId": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "runtimeMinutes": {
                    "type": "integer"
                },
                "seasonNumber": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "metadata.SearchHitResponse": {
            "type": "object",
            "properties": {
                "altTitles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "bannerUrl": {
                    "type": "string"
                },
                "endDate": {
                    "type": "string"
                },
                "episodeCount": {
                    "type": "integer"
                },
                "externalId": {
                    "type": "string"
                },
                "posterUrl": {
                    "type": "string"
                },
                "seasonCount": {
                    "type": "integer"
                },
                "startDate": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "synopsis": {
                    "type": "string"
                },
                "titleOriginal": {
                    "type": "string"
                },
                "titlePreferred": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "metadata.ShowResponse": {
            "type": "object",
            "properties": {
                "altTitles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "bannerUrl": {
                    "type": "string"
                },
                "endDate": {
                    "type": "string"
                },
                "episodeCount": {
                    "type": "integer"
                },
                "externalId": {
                    "type": "string"
                },
                "posterUrl": {
                    "type": "string"
                },
                "seasonCount": {
                    "type": "integer"
                },
                "startDate": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "synopsis": {
                    "type": "string"
                },
                "titleOriginal": {
                    "type": "string"
                },
                "titlePreferred": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
    }
}